De-dupe boilerplate test code
