+++ .github/workflows/build-test-push.yml
@@ -27,8 +27,6 @@ jobs:
         node-version: ${{ matrix.node-version }}
     - name: Install Dependencies
       run: yarn
-    - name: Remove OS Restrictions
-      run: node ./tasks/remove-os-restrictions.js
     - name: Upgrade Packages to Specified Closure Version
       run: node ./tasks/bump-version.js
     - name: Compile Native Binary
