+++ packages/google-closure-compiler-java/package.json
@@ -17,11 +17,5 @@
     "index.js",
     "package.json",
     "readme.md"
-  ],
-  "devDependencies": {
-    "kleur": "4.0.2"
-  },
-  "scripts": {
-    "test": "node ./test.js"
-  }
+  ]
 }
