--- packages/google-closure-compiler-java/test.js
@@ -1,28 +1,0 @@
-#!/usr/bin/env node
-/*
- * Copyright 2018 The Closure Compiler Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-const fs = require('fs');
-const javaPath = require('./');
-const kleur = require('kleur');
-
-process.stdout.write('google-closure-compiler-java\n');
-if (fs.existsSync(javaPath)) {
-  process.stdout.write(`  ${kleur.green('✓')} ${kleur.dim('compiler jar exists')}\n`);
-} else {
-  process.stdout.write(`  ${kleur.red('compiler jar does not exist')}\n`);
-  process.exitCode = 1;
-}
